+++ src/Owlia.Data/Migrations/20260828115414_AddSpeechPercentage.Designer.cs
@@ -2,6 +2,7 @@
 using System;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Owlia.Data;
 
@@ -10,9 +11,11 @@ using Owlia.Data;
 namespace Owlia.Data.Migrations
 {
     [DbContext(typeof(OwliaDbContext))]
-    partial class OwliaDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260828115414_AddSpeechPercentage")]
+    partial class AddSpeechPercentage
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder.HasAnnotation("ProductVersion", "10.0.11");